Removing some styles. Moved body of Credits, FAQ, Privacy, Terms Of Service into common files we can share across repositories.
