+++ src/index.html
@@ -2,7 +2,7 @@
 <!--
     index.html for We Vote Ballot
 -->
-<html lang='en'>
+<html lang="en">
   <head>
 <!--    <title><%= htmlWebpackPlugin.options.title %></title>-->
     <meta
