+++ src/js/components/Navigation/Header.jsx
@@ -163,9 +163,12 @@ export default class Header extends Component {
       path.startsWith('/about') ||
       path.startsWith('/for-campaigns') ||
       path.startsWith('/how/for-campaigns') ||
+      path.startsWith('/more/about') ||
+      path.startsWith('/more/credits') ||
       path.startsWith('/twitter_sign_in') ||
       path.startsWith('/wevoteintro') ||
-      path.startsWith('/welcome'));
+      path.startsWith('/welcome')
+    );
   }
 
 
@@ -439,7 +442,8 @@ export default class Header extends Component {
         </div>
       );
     } else if (typeof pathname !== 'undefined' && pathname &&
-      (pathname === '/for-campaigns' ||
+      (pathname === '/about' ||
+      pathname === '/for-campaigns' ||
       pathname === '/for-organizations' ||
       pathname.startsWith('/how') ||
       pathname === '/more/about' ||
