+++ src/js/routes/More/Credits.jsx
@@ -1,16 +1,12 @@
 import { withStyles } from '@material-ui/core/styles';
 import React, { Component, Suspense } from 'react';
 import Helmet from 'react-helmet';
-import { Link } from 'react-router-dom';
 import styled from 'styled-components';
-import { organizationalDonors, teamOfVolunteers } from '../../components/More/people';
+import CreditsBody from '../../common/components/CreditsBody';
 import { Section } from '../../components/Welcome/sectionStyles';
-import ExternalLinkIcon from '../../components/Widgets/ExternalLinkIcon';
 import cordovaScrollablePaneTopPadding from '../../utils/cordovaScrollablePaneTopPadding';
-import { isWebApp } from '../../utils/cordovaUtils';
 import { renderLog } from '../../utils/logging';
 
-const OpenExternalWebSite = React.lazy(() => import(/* webpackChunkName: 'OpenExternalWebSite' */ '../../components/Widgets/OpenExternalWebSite'));
 const WelcomeAppbar = React.lazy(() => import(/* webpackChunkName: 'WelcomeAppbar' */ '../../components/Navigation/WelcomeAppbar'));
 const WelcomeFooter = React.lazy(() => import(/* webpackChunkName: 'WelcomeFooter' */ '../../components/Welcome/WelcomeFooter'));
 
@@ -36,103 +32,7 @@ class Credits extends Component {
           <CreditsTitle>Credits &amp; Thanks</CreditsTitle>
         </HeaderForCredits>
         <Section noTopMargin>
-          <CreditsDescriptionContainer>
-            <span>
-              We are thankful to these organizations which are critical to our work.
-              {' '}
-              The companies on this list give us free or heavily discounted services (since we are a nonprofit), and
-              {' '}
-              the nonprofits overcome so many challenges to provide the data or other services we rely on.
-              {' '}
-              Please also see the
-              {' '}
-              <Link to="/more/attributions" id="attributions">
-                summary of open source software
-              </Link>
-              {' '}
-              We Vote uses.
-            </span>
-            <CompanyWrapper>
-              { organizationalDonors.map((item) => (
-                <div className="credits-company" key={item.alt}>
-                  <div>
-                    <div className="credits-company__logo-container">
-                      {
-                        item.logo && (
-                          <img className="credits-company__logo"
-                            src={item.logo}
-                            alt={`${item.alt} logo`}
-                          />
-                        )
-                      }
-                    </div>
-                    <div className="credits-company__info">
-                      {item.name && <strong>{item.name}</strong>}
-                      {item.title && (
-                        <p className="credits-company__title">
-                          {item.title}
-                        </p>
-                      )}
-                    </div>
-                  </div>
-                </div>
-              )) }
-            </CompanyWrapper>
-          </CreditsDescriptionContainer>
-          <CreditsDescriptionContainer>
-            <SectionTitle>Volunteers, Interns &amp; Donors</SectionTitle>
-            We couldn&apos;t do what we do without your help.
-            {' '}
-            Please join us by
-            <Suspense fallback={<></>}>
-              <OpenExternalWebSite
-                linkIdAttribute="wevoteJoinUs"
-                url="https://www.idealist.org/en/nonprofit/f917ce3db61a46cb8ad2b0d4e335f0af-we-vote-oakland#volops"
-                target="_blank"
-                className="open-web-site open-web-site__no-right-padding"
-                body={(
-                  <span>
-                    finding a role that excites you on our page at Idealist.org&nbsp;
-                    <ExternalLinkIcon />
-                  </span>
-                )}
-              />
-            </Suspense>
-            {isWebApp() && (
-              <span>
-                , or
-                {' '}
-                <Link to="/more/donate">
-                  donating now
-                </Link>
-              </span>
-            )}
-            .
-            <br />
-            <br />
-            <ul>
-              { teamOfVolunteers.map((item) => (
-                <div key={item.name}>
-                  <li>
-                    <strong>{item.name}</strong>
-                    {item.title && (
-                      <span>
-                        {' '}
-                        -
-                        {' '}
-                        {item.title}
-                      </span>
-                    )}
-                  </li>
-                </div>
-              ))}
-            </ul>
-            <br />
-            <br />
-            This list is in rough order of number of volunteer hours spent (10+ hours) or monetary donation level. Individual monetary donors only listed with express permission.
-            {' '}
-            (Our apologies if you should be on this list and are missing. Please contact Dale McGrew with corrections.)
-          </CreditsDescriptionContainer>
+          <CreditsBody />
         </Section>
         <Suspense fallback={<></>}>
           <WelcomeFooter />
@@ -194,41 +94,6 @@ const CreditsTitle = styled.h1`
   @media (max-width: ${({ theme }) => theme.breakpoints.xs}) {
     font-size: 18px;
     margin-top: 5em;
-  }
-`;
-
-// }
-const CreditsDescriptionContainer = styled.div`
-  margin: 1em auto;
-  width: 960px;
-  max-width: 90vw;
-  text-align: left;
-  @media (min-width: 960px) and (max-width: 991px) {
-    > * {
-      width: 90%;
-      margin: 0 auto;
-    }
-    max-width: 100%;
-    min-width: 100%;
-    width: 100%;
-    margin: 0 auto;
-  }
-`;
-
-const CompanyWrapper = styled.div`
-  display: flex;
-  flex-flow: row wrap;
-  justify-content: space-between;
-  text-align: center;
-`;
-
-const SectionTitle = styled.h1`
-  font-size: 28px;
-  font-weight: 300;
-  margin-bottom: 10px;
-  @media (max-width: ${({ theme }) => theme.breakpoints.md}) {
-    font-size: 24px;
-    margin-bottom: 9px;
   }
 `;
